requirements.txt
grpcio == 1.3.5
grpcio-tools == 1.3.5
protobuf>=3.2.0
leveldb == 0.193
coloredlogs
cryptography == 1.8.1
PyJWT
pyopenssl
yappi
Sphinx
sphinx_rtd_theme == 0.1.9
GitPython >= 2.1.1
tox
Flask
Flask-RESTful
requests